fix(ci): whitespace-insensitive ableset title fold + split state/mod.rs below size target
normalize_title_for_mismatch (#601) folded diacritics/punctuation/case but
kept internal whitespace significant, so "102 10000 armad" vs
"102 10 000 armád" — a diacritic-only pair that also happens to differ in
spacing — was classified as a genuine mismatch instead of silent formatting
variance, failing diacritic_only_difference_is_silent. Whitespace is the
same class of formatting noise as diacritics, so the normalizer now filters
to alphanumeric-only (dropping whitespace along with punctuation) before
lowercasing. The one test that asserted the old (now wrong) behavior,
normalize_title_for_mismatch_keeps_inner_whitespace_significant, is updated
to normalize_title_for_mismatch_treats_inner_whitespace_as_insignificant
with an assert_eq! and a comment explaining the design revision; every
other mismatch test already differs by more than whitespace and is
unaffected.

Also splits background-task spawning (heartbeat ticker, trash-prune loop,
timer/WAL/NDI-auto-reconnect tasks) out of state/mod.rs into a new sibling
module state/background_tasks.rs, following the #486 precedent. The
AbleSet batch's one-line `mod ableset_mismatch;` addition pushed mod.rs
from 799 to 868 prod lines, past the 800-line quality gate. Pure code
motion — no behavior change, no signature changes visible to callers
(methods stay on AppState, called the same way from new_with_heartbeat
and from_config). mod.rs is now 701 prod lines.
